+++ src/main/java/domain/Player.java
@@ -5,15 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class Player {
-    private BigInteger id;
-    private BigInteger teamId;
+    private Long id;
+    private Long teamId;
     private String name;
     private String position;
     private LocalDateTime createdAt;
